"steal" code from LLVM pertaining to uleb128 encoding
I don't like the look of the LLVM code one bit so I will integrate it
more fully into c-armyknife-lib so it seems native but I want to
preserve some history. My guess is that an LLM would now emit pretty
similar code to the llvm simply asking it to but how knows.