add buffer_begining_of_line and buffer_end_of_line
These are very similar (at least in spirit) to commands like C-a and C-e
in text editors like emacs. My plan is to use this in a tool to go back
several lines which I should be able to do using the function several
times plus a little addition or subtraction between calls to extract an
approcimate context around a point (as emacs calls it).