allow buffer's to retain their identity after growing
Previously the elements of a buffer always appear immediately after the
buffer but now we allocate (and reallocate) a separate array of bytes to
hold contents which means that buffer's themselves no longer move when
they grow. This makes them a little easier to use for many use cases
though potentially a little slower for certain use cases where the
buffer is kind of small as we may incur an additional cache miss since
the elements are no longer adjacent to the buffer header.