convert value-hashtable to using a seperately allocated array for the buckets
Previously the array lived immediately after the n_buckets and
n_entries. Like buffer_t this appears to not really be worth it versus
the pain that this causes (the goal is to have a smaller cache
footprint). This change means technically we can stop returning the
value hashtable on insert and delete.