we can now launch sub-processes and read their output
There is still room for improvement such as redirecting stdout and
stderr to the same pipe when we don't want to keep them seperate. We
also haven't implemented blocking or non-blocking writing to the
sub-processes stdin. Finally, a facility to write output(s) to a file or
read input from a file would be helpful. I think creating full shell
style pipelines is outside of the scope of this routine.

I highly doubt any of this would work under Windows (possibly even with
the Cosmopolitan C library?)