become more omni-c compatible at the "statement parse" level
Use cast() and compound_literal().

I still feel like cast() is an improvement overall - you get to see
where your code is *potentially* unsafe just by using grep (of course
not all uses of cast are unsafe, we can make this distinction later).

I'm not as certain about compound_literal. Shortening it to lit() is
possible and maybe making it not require the type would make it less
annoying but that will have to come later.